+++ firmware/application/ui_debug.hpp
@@ -130,42 +130,42 @@ private:
 };
 
 struct RegistersWidgetConfig {
-	int registers_count;
-	int register_bits;
+	size_t registers_count;
+	size_t register_bits;
 
-	constexpr int legend_length() const {
+	constexpr size_t legend_length() const {
 		return (registers_count >= 0x10) ? 2 : 1;
 	}
 
-	constexpr int legend_width() const {
+	constexpr size_t legend_width() const {
 		return legend_length() * 8;
 	}
 
-	constexpr int value_length() const {
+	constexpr size_t value_length() const {
 		return (register_bits + 3) / 4;
 	}
 
-	constexpr int value_width() const {
+	constexpr size_t value_width() const {
 		return value_length() * 8;
 	}
 
-	constexpr int registers_per_row() const {
+	constexpr size_t registers_per_row() const {
 		return (value_length() >= 3) ? 4 : 8;
 	}
 
-	constexpr int registers_row_length() const {
+	constexpr size_t registers_row_length() const {
 		return (registers_per_row() * (value_length() + 1)) - 1;
 	}
 
-	constexpr int registers_row_width() const {
+	constexpr size_t registers_row_width() const {
 		return registers_row_length() * 8;
 	}
 
-	constexpr int row_width() const {
+	constexpr size_t row_width() const {
 		return legend_width() + 8 + registers_row_width();
 	}
 
-	constexpr int rows() const {
+	constexpr size_t rows() const {
 		return registers_count / registers_per_row();
 	}
 };
@@ -185,7 +185,7 @@ private:
 	const RegistersWidgetConfig config;
 	const std::function<uint32_t(const size_t register_number)> reader;
 
-	static constexpr int row_height = 16;
+	static constexpr size_t row_height = 16;
 
 	void draw_legend(const Coord left, Painter& painter);
 	void draw_values(const Coord left, Painter& painter);
